+++ CHANGELOG.md
@@ -5,6 +5,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed 
+- Fixed a coding issue in the Push-EnrollUpdatableAsset function
+
 ## [1.0.17] - 2023-03-22
 ## Fixed
 - Fixed an issue with the Add-DeploymentAudienceMember function where it was not processing the input correctly.